move GHCSpecter.Util.Map to plugin:GHCSpecter.Data.Map.
